Refactor algorithm and analysis services
