+++ server/czcode_modules/clustering.py
@@ -1,2241 +1,17 @@
-import math
-
-import networkx as nx
-import numpy as np
-
-from common_geo import distance
-
-from .metrics import Helpers, cbg_population
-
-try:
-    from scipy.optimize import Bounds, LinearConstraint, milp
-    from scipy.sparse import coo_matrix
-    SCIPY_MILP_AVAILABLE = True
-except Exception:
-    SCIPY_MILP_AVAILABLE = False
+from .clustering_utils import ClusteringUtilsMixin
+from .directed_algorithms import DirectedClusteringMixin
+from .greedy_algorithms import GreedyClusteringMixin
+from .mobility_prune import MobilityPruneMixin
+from .optimal_czi import OptimalCziMixin
 
 
-class Clustering:
+class Clustering(
+    GreedyClusteringMixin,
+    DirectedClusteringMixin,
+    MobilityPruneMixin,
+    OptimalCziMixin,
+    ClusteringUtilsMixin,
+):
     def __init__(self, config, logger):
         self.config = config
         self.logger = logger
-
-    @staticmethod
-    def _record_trace_step(
-        trace_collector,
-        iteration,
-        cluster_before,
-        population_before,
-        candidates,
-        selected_cbg,
-        selected_population,
-        cluster_after,
-        population_after,
-        metrics_after=None,
-        higher_score_better=True
-    ):
-        if trace_collector is None:
-            return
-
-        sorted_candidates = sorted(
-            candidates,
-            key=lambda item: (float(item.get('score', 0.0)), item.get('cbg', '')),
-            reverse=bool(higher_score_better)
-        )
-        for idx, candidate in enumerate(sorted_candidates):
-            candidate['rank'] = idx + 1
-            candidate['selected'] = (candidate.get('cbg') == selected_cbg)
-
-        step = {
-            'iteration': int(iteration),
-            'cluster_before': list(cluster_before),
-            'population_before': int(population_before),
-            'candidates': sorted_candidates,
-            'selected_cbg': selected_cbg,
-            'selected_population': int(selected_population),
-            'cluster_after': list(cluster_after),
-            'population_after': int(population_after),
-        }
-        if metrics_after:
-            step['metrics_after'] = metrics_after
-        trace_collector.append(step)
-
-    @staticmethod
-    def _movement_outside_cluster(G: nx.Graph, candidate, cluster_set):
-        movement_out = 0.0
-        if candidate not in G:
-            return movement_out
-        for neighbor in G.adj[candidate]:
-            if neighbor in cluster_set:
-                continue
-            movement_out += float(G.adj[candidate][neighbor].get('weight', 0))
-        return movement_out
-
-    @staticmethod
-    def _seed_distance_km(seed_cbg, candidate_cbg, cbg_centers):
-        if not cbg_centers:
-            return None
-
-        seed_center = cbg_centers.get(seed_cbg)
-        cand_center = cbg_centers.get(candidate_cbg)
-        if seed_center is None or cand_center is None:
-            return None
-
-        try:
-            return float(distance(
-                seed_center[0],
-                seed_center[1],
-                cand_center[0],
-                cand_center[1],
-            ))
-        except Exception:
-            return None
-
-    @classmethod
-    def _seed_guard_membership(cls, seed_cbg, candidate_cbg, cbg_centers, seed_guard_distance_km):
-        if candidate_cbg == seed_cbg:
-            return True, 0.0
-
-        seed_distance = cls._seed_distance_km(seed_cbg, candidate_cbg, cbg_centers)
-        try:
-            threshold_km = float(seed_guard_distance_km)
-        except (TypeError, ValueError):
-            threshold_km = 0.0
-
-        if threshold_km <= 0 or seed_distance is None:
-            return True, seed_distance
-        return seed_distance <= threshold_km, seed_distance
-
-    @classmethod
-    def _seed_guard_contributor_sets(cls, seed_cbg, cluster_cbgs, cbg_centers, seed_guard_distance_km):
-        contributor_set = set()
-        excluded_set = set()
-
-        for cbg in cluster_cbgs:
-            contributes, _ = cls._seed_guard_membership(
-                seed_cbg,
-                cbg,
-                cbg_centers,
-                seed_guard_distance_km,
-            )
-            if contributes:
-                contributor_set.add(cbg)
-            else:
-                excluded_set.add(cbg)
-
-        if seed_cbg in cluster_cbgs:
-            contributor_set.add(seed_cbg)
-            excluded_set.discard(seed_cbg)
-
-        return contributor_set, excluded_set
-
-    @staticmethod
-    def _seed_centroid(seed_cbgs, cbg_centers):
-        if not cbg_centers:
-            return None
-
-        coords = []
-        for cbg in seed_cbgs or []:
-            center = cbg_centers.get(cbg)
-            if center is None:
-                continue
-            coords.append((float(center[0]), float(center[1])))
-
-        if not coords:
-            return None
-
-        lat = sum(item[0] for item in coords) / len(coords)
-        lon = sum(item[1] for item in coords) / len(coords)
-        return (lat, lon)
-
-    @staticmethod
-    def _directed_containment(DG, cluster_set):
-        internal = 0.0
-        out_origin = 0.0
-        out_dest = 0.0
-
-        for node in cluster_set:
-            if node not in DG:
-                continue
-
-            internal += float(DG.nodes[node].get('self_weight', 0) or 0)
-
-            for _, v, data in DG.out_edges(node, data=True):
-                if v == node:
-                    continue
-                w = float(data.get('weight', 0) or 0)
-                if v in cluster_set:
-                    internal += w
-                else:
-                    out_origin += w
-
-            for u, _, data in DG.in_edges(node, data=True):
-                if u == node or u in cluster_set:
-                    continue
-                out_dest += float(data.get('weight', 0) or 0)
-
-        orig_total = internal + out_origin
-        dest_total = internal + out_dest
-        origin_containment = internal / orig_total if orig_total > 0 else 0.0
-        destination_containment = internal / dest_total if dest_total > 0 else 0.0
-        zone_containment = min(origin_containment, destination_containment)
-
-        return {
-            'internal': internal,
-            'out_origin': out_origin,
-            'out_dest': out_dest,
-            'origin': origin_containment,
-            'destination': destination_containment,
-            'zone': zone_containment,
-        }
-
-    @staticmethod
-    def _within_local_radius(seed_centroid, candidate_cbg, cbg_centers, local_radius_km):
-        if seed_centroid is None or not cbg_centers:
-            return True
-
-        try:
-            radius_km = float(local_radius_km)
-        except (TypeError, ValueError):
-            radius_km = 0.0
-
-        if radius_km <= 0:
-            return True
-
-        candidate_center = cbg_centers.get(candidate_cbg)
-        if candidate_center is None:
-            return False
-
-        try:
-            dist_km = float(distance(
-                seed_centroid[0],
-                seed_centroid[1],
-                candidate_center[0],
-                candidate_center[1],
-            ))
-        except Exception:
-            return False
-
-        return dist_km <= radius_km
-
-    @staticmethod
-    def _node_total_directed_flow(DG, node):
-        if node not in DG:
-            return 0.0
-
-        total = float(DG.nodes[node].get('self_weight', 0) or 0)
-        for _, v, data in DG.out_edges(node, data=True):
-            if v == node:
-                continue
-            total += float(data.get('weight', 0) or 0)
-        for u, _, data in DG.in_edges(node, data=True):
-            if u == node:
-                continue
-            total += float(data.get('weight', 0) or 0)
-        return total
-
-    def greedy_fast(self, G: nx.Graph, u0: str, min_pop: int, trace_collector=None):
-        self.logger.info(f"Starting greedy_fast algorithm with seed CBG {u0}")
-        population = cbg_population(u0, self.config, self.logger)
-
-        cluster = [u0]
-        cluster_set = {u0}
-        surround = list(set([j for j in list(G.adj[u0]) if j not in cluster_set]))
-
-        for j in list(G.adj[u0]):
-            if j not in surround and j not in cluster_set:
-                surround.append(j)
-
-        if len(surround) == 0:
-            self.logger.warning("No adjacent CBGs found. Cannot reach target population.")
-            return cluster, population
-
-        itr = 0
-
-        while population < min_pop:
-            max_weight = 0
-            best_cbg = surround[0]
-            best_pop = 0
-            candidate_details = []
-
-            for candidate in surround:
-                if candidate in cluster_set:
-                    continue
-
-                cur_pop = cbg_population(candidate, self.config, self.logger)
-                if cur_pop == 0:
-                    continue
-
-                weight = sum([G.get_edge_data(candidate, cbg, {}).get('weight', 0) for cbg in cluster])
-                weight = float(weight)
-                movement_outside = self._movement_outside_cluster(G, candidate, cluster_set)
-                candidate_details.append({
-                    'cbg': candidate,
-                    'population': int(cur_pop),
-                    'score': weight,
-                    'movement_to_cluster': weight,
-                    'movement_to_outside': movement_outside,
-                })
-
-                if weight > max_weight:
-                    max_weight = weight
-                    best_cbg = candidate
-                    best_pop = cur_pop
-
-            prev_cluster = list(cluster)
-            prev_population = population
-
-            surround.remove(best_cbg)
-            cluster.append(best_cbg)
-            cluster_set.add(best_cbg)
-            population += best_pop
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=itr,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=candidate_details,
-                selected_cbg=best_cbg,
-                selected_population=best_pop,
-                cluster_after=cluster,
-                population_after=population
-            )
-
-            self.logger.info(f"Iteration {itr}: Added CBG {best_cbg} with pop {best_pop}. New total: {population}")
-
-            surround.extend([j for j in list(G.adj[best_cbg]) if j not in cluster_set])
-            surround = list(set(surround))
-
-            itr += 1
-            if itr > 500:
-                self.logger.warning("Max iterations exceeded (500). Cannot reach target population.")
-                break
-
-        return cluster, population
-
-    def greedy_weight(self, G, u0, min_pop, trace_collector=None):
-        self.logger.info(f"Starting greedy_weight algorithm with seed CBG {u0}")
-        cluster = [u0]
-        cluster_set = {u0}
-        population = cbg_population(u0, self.config, self.logger)
-        it = 1
-        self.logger.info(f"Seed CBG population: {population}")
-        while population < min_pop:
-            all_adj_cbgs = []
-            for i in cluster:
-                try:
-                    for j in list(G.adj[i]):
-                        if j not in all_adj_cbgs and j not in cluster_set:
-                            all_adj_cbgs.append(j)
-                except KeyError:
-                    self.logger.warning(f"CBG {i} not found in graph")
-                    continue
-            if not all_adj_cbgs:
-                self.logger.warning(f"No adjacent CBGs found after {it} iterations. Cannot reach target population.")
-                break
-            max_movement = 0
-            cbg_to_add = all_adj_cbgs[0]
-            candidate_details = []
-            for candidate in all_adj_cbgs:
-                current_movement = 0
-                for member in cluster:
-                    try:
-                        current_movement += G.adj[candidate][member]['weight']
-                    except (KeyError, ZeroDivisionError):
-                        continue
-                movement_to_cluster = float(current_movement)
-                movement_to_outside = self._movement_outside_cluster(G, candidate, cluster_set)
-                candidate_details.append({
-                    'cbg': candidate,
-                    'population': int(cbg_population(candidate, self.config, self.logger)),
-                    'score': movement_to_cluster,
-                    'movement_to_cluster': movement_to_cluster,
-                    'movement_to_outside': movement_to_outside,
-                })
-                if current_movement > max_movement:
-                    max_movement = current_movement
-                    cbg_to_add = candidate
-
-            prev_cluster = list(cluster)
-            prev_population = population
-            cluster.append(cbg_to_add)
-            cluster_set.add(cbg_to_add)
-            cbg_pop = cbg_population(cbg_to_add, self.config, self.logger)
-            population += cbg_pop
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=it - 1,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=candidate_details,
-                selected_cbg=cbg_to_add,
-                selected_population=cbg_pop,
-                cluster_after=cluster,
-                population_after=population
-            )
-
-            self.logger.info(f"Iteration {it}: Added CBG {cbg_to_add} with pop {cbg_pop}. New total: {population}")
-            it += 1
-            if it > 1000:
-                self.logger.warning("Reached maximum iterations (1000). Stopping algorithm.")
-                break
-        return cluster, population
-
-    def greedy_weight_seed_guard(
-        self,
-        G,
-        u0,
-        min_pop,
-        seed_guard_distance_km=20.0,
-        cbg_centers=None,
-        trace_collector=None,
-    ):
-        self.logger.info(
-            "Starting greedy_weight_seed_guard algorithm with seed CBG %s (distance %.2f km)",
-            u0,
-            float(seed_guard_distance_km),
-        )
-        cluster = [u0]
-        cluster_set = {u0}
-        contributor_set = {u0}
-        excluded_set = set()
-        population = cbg_population(u0, self.config, self.logger)
-        cbg_centers = cbg_centers or {}
-        it = 1
-
-        self.logger.info(f"Seed CBG population: {population}")
-        while population < min_pop:
-            all_adj_cbgs = []
-            for i in cluster:
-                try:
-                    for j in list(G.adj[i]):
-                        if j not in all_adj_cbgs and j not in cluster_set:
-                            all_adj_cbgs.append(j)
-                except KeyError:
-                    self.logger.warning(f"CBG {i} not found in graph")
-                    continue
-
-            if not all_adj_cbgs:
-                self.logger.warning(
-                    f"No adjacent CBGs found after {it} iterations. Cannot reach target population."
-                )
-                break
-
-            best_choice = None
-            candidate_details = []
-            for candidate in all_adj_cbgs:
-                movement_to_cluster = 0.0
-                movement_to_full_cluster = 0.0
-                for member in cluster:
-                    try:
-                        edge_weight = float(G.adj[candidate][member]['weight'])
-                    except (KeyError, ZeroDivisionError):
-                        continue
-                    movement_to_full_cluster += edge_weight
-                    if member in contributor_set:
-                        movement_to_cluster += edge_weight
-
-                movement_to_outside = self._movement_outside_cluster(G, candidate, cluster_set)
-                candidate_pop = int(cbg_population(candidate, self.config, self.logger))
-                contributes_after, seed_distance = self._seed_guard_membership(
-                    u0,
-                    candidate,
-                    cbg_centers,
-                    seed_guard_distance_km,
-                )
-                distance_tiebreak = 0.0 if seed_distance is None else -float(seed_distance)
-                candidate_tuple = (
-                    float(movement_to_cluster),
-                    int(contributes_after),
-                    distance_tiebreak,
-                    float(candidate_pop),
-                    str(candidate),
-                )
-                candidate_details.append({
-                    'cbg': candidate,
-                    'population': candidate_pop,
-                    'score': float(movement_to_cluster),
-                    'movement_to_cluster': float(movement_to_cluster),
-                    'movement_to_full_cluster': float(movement_to_full_cluster),
-                    'movement_to_outside': float(movement_to_outside),
-                    'seed_distance_km': (
-                        float(seed_distance) if seed_distance is not None else None
-                    ),
-                    'movement_contributes_after_selection': bool(contributes_after),
-                })
-                if best_choice is None or candidate_tuple > best_choice[0]:
-                    best_choice = (
-                        candidate_tuple,
-                        candidate,
-                        candidate_pop,
-                        contributes_after,
-                        seed_distance,
-                    )
-
-            if best_choice is None:
-                self.logger.warning(
-                    f"No valid candidate CBGs found after {it} iterations. Cannot reach target population."
-                )
-                break
-
-            _, cbg_to_add, cbg_pop, contributes_after, seed_distance = best_choice
-            prev_cluster = list(cluster)
-            prev_population = population
-            cluster.append(cbg_to_add)
-            cluster_set.add(cbg_to_add)
-            population += cbg_pop
-
-            if contributes_after:
-                contributor_set.add(cbg_to_add)
-            else:
-                excluded_set.add(cbg_to_add)
-
-            metrics_after = {
-                'seed_guard_distance_km': float(seed_guard_distance_km),
-                'movement_contributor_count': len(contributor_set),
-                'movement_excluded_count': len(excluded_set),
-            }
-            if excluded_set:
-                metrics_after['movement_excluded_cbgs'] = list(excluded_set)
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=it - 1,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=candidate_details,
-                selected_cbg=cbg_to_add,
-                selected_population=cbg_pop,
-                cluster_after=cluster,
-                population_after=population,
-                metrics_after=metrics_after,
-            )
-
-            movement_note = "counts toward future scoring"
-            if not contributes_after:
-                movement_note = "excluded from future scoring"
-            self.logger.info(
-                "Iteration %d: Added CBG %s with pop %d. New total: %d (%s, distance=%s km)",
-                it,
-                cbg_to_add,
-                cbg_pop,
-                population,
-                movement_note,
-                (
-                    f"{float(seed_distance):.2f}"
-                    if seed_distance is not None
-                    else "unknown"
-                ),
-            )
-            it += 1
-            if it > 1000:
-                self.logger.warning("Reached maximum iterations (1000). Stopping algorithm.")
-                break
-        return cluster, population
-
-    def greedy_ratio(self, G, u0, min_pop, trace_collector=None):
-        self.logger.info(f"Starting greedy_ratio algorithm with seed CBG {u0}")
-        cluster = [u0]
-        cluster_set = {u0}
-        population = cbg_population(u0, self.config, self.logger)
-        it = 1
-        self.logger.info(f"Seed CBG population: {population}")
-        while population < min_pop:
-            all_adj_cbgs = []
-            for i in cluster:
-                try:
-                    for j in list(G.adj[i]):
-                        if j not in all_adj_cbgs and j not in cluster_set:
-                            all_adj_cbgs.append(j)
-                except KeyError:
-                    self.logger.warning(f"CBG {i} not found in graph")
-                    continue
-            if not all_adj_cbgs:
-                self.logger.warning(f"No adjacent CBGs found after {it} iterations. Cannot reach target population.")
-                break
-            max_ratio = 0
-            cbg_to_add = all_adj_cbgs[0]
-            candidate_details = []
-            for candidate in all_adj_cbgs:
-                movement_in = 0
-                movement_out = 0
-                for j in G.adj[candidate]:
-                    if j in cluster:
-                        movement_in += G.adj[candidate][j]['weight']
-                    else:
-                        movement_out += G.adj[candidate][j]['weight']
-                total_movement = movement_in + movement_out
-                ratio = 0.0
-                if total_movement > 0:
-                    ratio = movement_in / total_movement
-                    if ratio > max_ratio:
-                        max_ratio = ratio
-                        cbg_to_add = candidate
-                candidate_details.append({
-                    'cbg': candidate,
-                    'population': int(cbg_population(candidate, self.config, self.logger)),
-                    'score': float(ratio),
-                    'movement_to_cluster': float(movement_in),
-                    'movement_to_outside': float(movement_out),
-                    'movement_total': float(total_movement),
-                })
-
-            prev_cluster = list(cluster)
-            prev_population = population
-            cluster.append(cbg_to_add)
-            cluster_set.add(cbg_to_add)
-            cbg_pop = cbg_population(cbg_to_add, self.config, self.logger)
-            population += cbg_pop
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=it - 1,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=candidate_details,
-                selected_cbg=cbg_to_add,
-                selected_population=cbg_pop,
-                cluster_after=cluster,
-                population_after=population
-            )
-
-            self.logger.info(f"Iteration {it}: Added CBG {cbg_to_add} with pop {cbg_pop}. New total: {population}")
-            it += 1
-            if it > 1000:
-                self.logger.warning("Reached maximum iterations (1000). Stopping algorithm.")
-                break
-        return cluster, population
-
-    def greedy_ttwa(
-        self,
-        DG,
-        u0,
-        min_pop,
-        containment_threshold=0.70,
-        max_iter=1000,
-        trace_collector=None,
-    ):
-        self.logger.info(
-            f"Starting greedy_ttwa with seed CBG {u0} "
-            f"(threshold={containment_threshold}, min_pop={min_pop})"
-        )
-
-        if u0 not in DG:
-            raise ValueError(f"Seed CBG {u0} not present in directed mobility graph")
-
-        def _self_weight(node):
-            return float(DG.nodes[node].get('self_weight', 0) or 0)
-
-        cluster = [u0]
-        cluster_set = {u0}
-        population = cbg_population(u0, self.config, self.logger)
-
-        internal = _self_weight(u0)
-        out_origin = 0.0
-        out_dest = 0.0
-        for _, v, data in DG.out_edges(u0, data=True):
-            if v == u0:
-                continue
-            out_origin += float(data.get('weight', 0))
-        for u, _, data in DG.in_edges(u0, data=True):
-            if u == u0:
-                continue
-            out_dest += float(data.get('weight', 0))
-
-        def _containment(internal_, out_origin_, out_dest_):
-            orig_total = internal_ + out_origin_
-            dest_total = internal_ + out_dest_
-            orig_c = internal_ / orig_total if orig_total > 0 else 0.0
-            dest_c = internal_ / dest_total if dest_total > 0 else 0.0
-            return orig_c, dest_c, min(orig_c, dest_c)
-
-        orig_c, dest_c, zone_c = _containment(internal, out_origin, out_dest)
-        self.logger.info(
-            f"Seed containment: origin={orig_c:.4f} dest={dest_c:.4f} "
-            f"zone={zone_c:.4f} pop={population}"
-        )
-
-        it = 1
-        while it <= max_iter:
-            candidates = set()
-            for node in cluster:
-                candidates.update(DG.successors(node))
-                candidates.update(DG.predecessors(node))
-            candidates -= cluster_set
-            if not candidates:
-                self.logger.info("No further candidates; stopping.")
-                break
-
-            best = None
-            best_score = None
-            best_deltas = None
-            candidate_details = []
-            for cand in candidates:
-                delta_internal = _self_weight(cand)
-                delta_out_origin = 0.0
-                delta_out_dest = 0.0
-
-                for _, y, data in DG.out_edges(cand, data=True):
-                    if y == cand:
-                        continue
-                    w = float(data.get('weight', 0))
-                    if y in cluster_set:
-                        delta_internal += w
-                        delta_out_dest -= w
-                    else:
-                        delta_out_origin += w
-
-                for y, _, data in DG.in_edges(cand, data=True):
-                    if y == cand:
-                        continue
-                    w = float(data.get('weight', 0))
-                    if y in cluster_set:
-                        delta_internal += w
-                        delta_out_origin -= w
-                    else:
-                        delta_out_dest += w
-
-                new_internal = internal + delta_internal
-                new_out_origin = out_origin + delta_out_origin
-                new_out_dest = out_dest + delta_out_dest
-                _, _, new_zone_c = _containment(new_internal, new_out_origin, new_out_dest)
-                score = new_zone_c - zone_c
-
-                candidate_details.append({
-                    'cbg': cand,
-                    'population': int(cbg_population(cand, self.config, self.logger)),
-                    'score': float(score),
-                    'new_zone_containment': float(new_zone_c),
-                    'delta_internal': float(delta_internal),
-                    'delta_out_origin': float(delta_out_origin),
-                    'delta_out_dest': float(delta_out_dest),
-                })
-
-                if (
-                    best_score is None
-                    or score > best_score
-                    or (score == best_score and (best is None or cand < best))
-                ):
-                    best_score = score
-                    best = cand
-                    best_deltas = (delta_internal, delta_out_origin, delta_out_dest)
-
-            if best is None:
-                break
-
-            if population >= min_pop and best_score <= 0:
-                self.logger.info(
-                    f"No candidate improves zone containment (best Δ={best_score:.4g}); "
-                    f"stopping with zone_c={zone_c:.4f} pop={population}"
-                )
-                break
-
-            chosen_pop = cbg_population(best, self.config, self.logger)
-            prev_cluster = list(cluster)
-            prev_population = population
-
-            cluster.append(best)
-            cluster_set.add(best)
-            d_int, d_oo, d_od = best_deltas
-            internal += d_int
-            out_origin += d_oo
-            out_dest += d_od
-            orig_c, dest_c, zone_c = _containment(internal, out_origin, out_dest)
-            population += chosen_pop
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=it - 1,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=candidate_details,
-                selected_cbg=best,
-                selected_population=chosen_pop,
-                cluster_after=cluster,
-                population_after=population,
-            )
-
-            self.logger.info(
-                f"Iteration {it}: added {best} (pop {chosen_pop}); "
-                f"zone_c={zone_c:.4f} (orig={orig_c:.4f}, dest={dest_c:.4f}) "
-                f"pop={population}"
-            )
-            it += 1
-
-        if it > max_iter:
-            self.logger.warning(f"Reached max_iter={max_iter}; stopping.")
-
-        if zone_c < containment_threshold:
-            self.logger.warning(
-                f"Final zone_c={zone_c:.4f} below threshold "
-                f"{containment_threshold}; zone may be weakly self-contained."
-            )
-
-        return cluster, population
-
-    def hierarchical_core_satellites(
-        self,
-        DG,
-        seed_cbgs,
-        min_pop,
-        cbg_to_zip=None,
-        zip_to_cbgs=None,
-        cbg_centers=None,
-        local_radius_km=20.0,
-        core_containment_threshold=0.70,
-        core_improvement_epsilon=0.0025,
-        satellite_flow_threshold=0.02,
-        max_satellites=4,
-        max_iter=1000,
-        trace_collector=None,
-    ):
-        cbg_to_zip = cbg_to_zip or {}
-        zip_to_cbgs = zip_to_cbgs or {}
-        cbg_centers = cbg_centers or {}
-
-        seed_cluster = []
-        missing_seed_cbgs = []
-        seen = set()
-        for cbg in seed_cbgs or []:
-            if cbg in seen:
-                continue
-            seen.add(cbg)
-            if cbg in DG:
-                seed_cluster.append(cbg)
-            else:
-                missing_seed_cbgs.append(cbg)
-
-        if not seed_cluster:
-            raise ValueError("None of the seed region CBGs are present in the directed mobility graph")
-
-        self.logger.info(
-            "Starting hierarchical_core_satellites with %d seed CBGs, min_pop=%s, local_radius_km=%s",
-            len(seed_cluster),
-            int(min_pop),
-            float(local_radius_km),
-        )
-
-        core_cluster = list(seed_cluster)
-        core_set = set(core_cluster)
-        seed_centroid = self._seed_centroid(seed_cluster, cbg_centers)
-        seed_zip_codes = sorted({
-            zip_code for zip_code in (cbg_to_zip.get(cbg) for cbg in seed_cluster) if zip_code
-        })
-
-        core_population = sum(cbg_population(cbg, self.config, self.logger) for cbg in core_cluster)
-        core_metrics = self._directed_containment(DG, core_set)
-
-        self.logger.info(
-            "Initial seed region: pop=%d origin=%.4f dest=%.4f zone=%.4f",
-            core_population,
-            core_metrics['origin'],
-            core_metrics['destination'],
-            core_metrics['zone'],
-        )
-
-        iteration = 1
-        while iteration <= max_iter:
-            local_candidates = set()
-            for node in core_cluster:
-                if node not in DG:
-                    continue
-                for candidate in DG.successors(node):
-                    if candidate in core_set:
-                        continue
-                    if not self._within_local_radius(seed_centroid, candidate, cbg_centers, local_radius_km):
-                        continue
-                    local_candidates.add(candidate)
-                for candidate in DG.predecessors(node):
-                    if candidate in core_set:
-                        continue
-                    if not self._within_local_radius(seed_centroid, candidate, cbg_centers, local_radius_km):
-                        continue
-                    local_candidates.add(candidate)
-
-            if not local_candidates:
-                self.logger.info("No local-core candidates remain; stopping local growth.")
-                break
-
-            best_candidate = None
-            best_metrics = None
-            best_score = None
-            candidate_details = []
-
-            for candidate in sorted(local_candidates):
-                next_set = set(core_set)
-                next_set.add(candidate)
-                next_metrics = self._directed_containment(DG, next_set)
-                score = float(next_metrics['zone'] - core_metrics['zone'])
-                candidate_zip = cbg_to_zip.get(candidate)
-                candidate_details.append({
-                    'cbg': candidate,
-                    'population': int(cbg_population(candidate, self.config, self.logger)),
-                    'score': score,
-                    'new_zone_containment': float(next_metrics['zone']),
-                    'origin_containment': float(next_metrics['origin']),
-                    'destination_containment': float(next_metrics['destination']),
-                    'seed_zip': candidate_zip,
-                })
-
-                if (
-                    best_score is None
-                    or score > best_score
-                    or (
-                        abs(score - best_score) < 1e-12
-                        and (best_candidate is None or candidate < best_candidate)
-                    )
-                ):
-                    best_candidate = candidate
-                    best_metrics = next_metrics
-                    best_score = score
-
-            if best_candidate is None:
-                break
-
-            if best_score <= 0:
-                self.logger.info(
-                    "Local core stabilized with no improving candidate (best delta=%.4g).",
-                    float(best_score),
-                )
-                break
-
-            if (
-                core_metrics['zone'] >= float(core_containment_threshold)
-                and best_score <= float(core_improvement_epsilon)
-            ):
-                self.logger.info(
-                    "Local core reached containment target %.4f with marginal gain %.4g; stopping.",
-                    float(core_metrics['zone']),
-                    float(best_score),
-                )
-                break
-
-            chosen_population = int(cbg_population(best_candidate, self.config, self.logger))
-            prev_cluster = list(core_cluster)
-            prev_population = int(core_population)
-
-            core_cluster.append(best_candidate)
-            core_set.add(best_candidate)
-            core_population += chosen_population
-            core_metrics = best_metrics
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=iteration - 1,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=candidate_details,
-                selected_cbg=best_candidate,
-                selected_population=chosen_population,
-                cluster_after=core_cluster,
-                population_after=core_population,
-                metrics_after={
-                    'stage': 'local_core',
-                    'origin_containment': float(core_metrics['origin']),
-                    'destination_containment': float(core_metrics['destination']),
-                    'zone_containment': float(core_metrics['zone']),
-                },
-            )
-
-            self.logger.info(
-                "Local core iteration %d: added %s pop=%d zone=%.4f",
-                iteration,
-                best_candidate,
-                chosen_population,
-                core_metrics['zone'],
-            )
-            iteration += 1
-
-        final_cluster = list(core_cluster)
-        final_set = set(core_set)
-        total_population = int(core_population)
-
-        satellite_candidates = []
-        satellite_units = {}
-        for node in core_cluster:
-            if node not in DG:
-                continue
-            neighbor_iter = list(DG.successors(node)) + list(DG.predecessors(node))
-            for candidate in neighbor_iter:
-                if candidate in final_set:
-                    continue
-                unit_id = cbg_to_zip.get(candidate)
-                if not unit_id or unit_id in seed_zip_codes:
-                    continue
-                satellite_units.setdefault(unit_id, set())
-
-        for unit_id in list(satellite_units.keys()):
-            unit_cbgs = []
-            for cbg in zip_to_cbgs.get(unit_id, []):
-                if cbg in final_set or cbg not in DG:
-                    continue
-                unit_cbgs.append(cbg)
-
-            if not unit_cbgs:
-                continue
-
-            shared_flow = 0.0
-            total_flow = 0.0
-            unit_population = 0
-            for cbg in unit_cbgs:
-                unit_population += int(cbg_population(cbg, self.config, self.logger))
-                total_flow += self._node_total_directed_flow(DG, cbg)
-
-                for _, v, data in DG.out_edges(cbg, data=True):
-                    if v in core_set:
-                        shared_flow += float(data.get('weight', 0) or 0)
-                for u, _, data in DG.in_edges(cbg, data=True):
-                    if u in core_set:
-                        shared_flow += float(data.get('weight', 0) or 0)
-
-            if unit_population <= 0 or shared_flow <= 0 or total_flow <= 0:
-                continue
-
-            coupling = float(shared_flow / total_flow)
-            satellite_candidates.append({
-                'unit_id': unit_id,
-                'label': f"ZIP {unit_id}",
-                'cbgs': sorted(unit_cbgs),
-                'population': int(unit_population),
-                'shared_flow': float(shared_flow),
-                'total_flow': float(total_flow),
-                'coupling': coupling,
-                'cbg_count': len(unit_cbgs),
-            })
-
-        satellite_candidates.sort(
-            key=lambda item: (
-                float(item['coupling']),
-                float(item['shared_flow']),
-                int(item['population']),
-                str(item['unit_id']),
-            ),
-            reverse=True,
-        )
-
-        selected_satellites = []
-        try:
-            max_satellites_int = max(0, int(max_satellites))
-        except (TypeError, ValueError):
-            max_satellites_int = 0
-
-        for candidate in satellite_candidates:
-            if max_satellites_int == 0:
-                break
-            if max_satellites_int and len(selected_satellites) >= max_satellites_int:
-                break
-            if candidate['coupling'] <= 0:
-                break
-            if total_population >= int(min_pop) and candidate['coupling'] < float(satellite_flow_threshold):
-                break
-
-            added_cbgs = [cbg for cbg in candidate['cbgs'] if cbg not in final_set]
-            if not added_cbgs:
-                continue
-
-            final_cluster.extend(added_cbgs)
-            final_set.update(added_cbgs)
-            total_population += int(candidate['population'])
-            selected_satellites.append({
-                'unit_id': candidate['unit_id'],
-                'label': candidate['label'],
-                'population': int(candidate['population']),
-                'coupling': float(candidate['coupling']),
-                'shared_flow': float(candidate['shared_flow']),
-                'cbg_count': int(candidate['cbg_count']),
-            })
-
-        final_metrics = self._directed_containment(DG, final_set)
-        metadata = {
-            'seed_cbgs': list(seed_cluster),
-            'missing_seed_cbgs': list(missing_seed_cbgs),
-            'seed_zip_codes': seed_zip_codes,
-            'core_cluster': list(core_cluster),
-            'core_population': int(core_population),
-            'core_containment': {
-                'origin': float(core_metrics['origin']),
-                'destination': float(core_metrics['destination']),
-                'zone': float(core_metrics['zone']),
-            },
-            'selected_satellites': selected_satellites,
-            'satellite_candidate_count': len(satellite_candidates),
-            'final_containment': {
-                'origin': float(final_metrics['origin']),
-                'destination': float(final_metrics['destination']),
-                'zone': float(final_metrics['zone']),
-            },
-            'external_pressure_share': float(max(0.0, 1.0 - final_metrics['zone'])),
-            'population_target_met': bool(total_population >= int(min_pop)),
-        }
-
-        if total_population < int(min_pop):
-            self.logger.warning(
-                "Hierarchical cluster ended below target population: pop=%d target=%d",
-                total_population,
-                int(min_pop),
-            )
-
-        return final_cluster, total_population, metadata
-
-    def mobility_prune(
-        self,
-        G: nx.Graph,
-        seed_cbgs,
-        min_pop: int,
-        max_iter: int = 1000,
-        trace_collector=None,
-        envelope_population_multiplier: float = 2.0,
-        envelope_population_floor: int = 100000,
-        envelope_max_cbgs: int = 250,
-        min_seed_capture: float = 0.80,
-        trace_candidate_limit: int = 50,
-    ):
-        seed_cluster = []
-        missing_seed_cbgs = []
-        seen = set()
-        for cbg in seed_cbgs or []:
-            if cbg in seen:
-                continue
-            seen.add(cbg)
-            if cbg in G:
-                seed_cluster.append(cbg)
-            else:
-                missing_seed_cbgs.append(cbg)
-
-        if not seed_cluster:
-            raise ValueError("None of the seed region CBGs are present in the mobility graph")
-
-        self.logger.info(
-            "Starting mobility_prune with %d seed CBGs",
-            len(seed_cluster),
-        )
-
-        seed_set = set(seed_cluster)
-        cluster = list(seed_cluster)
-        cluster_set = set(cluster)
-        population_by_cbg = {}
-
-        def get_population(cbg):
-            if cbg not in population_by_cbg:
-                population_by_cbg[cbg] = max(
-                    0,
-                    int(cbg_population(cbg, self.config, self.logger) or 0),
-                )
-            return population_by_cbg[cbg]
-
-        population = sum(get_population(cbg) for cbg in cluster)
-        seed_population = int(population)
-
-        try:
-            legacy_min_population = max(0, int(min_pop))
-        except (TypeError, ValueError):
-            legacy_min_population = 0
-
-        try:
-            multiplier = float(envelope_population_multiplier)
-        except (TypeError, ValueError):
-            multiplier = 2.0
-        if multiplier < 1.0:
-            multiplier = 1.0
-
-        try:
-            population_floor = max(0, int(envelope_population_floor))
-        except (TypeError, ValueError):
-            population_floor = 100000
-
-        envelope_population_target = max(
-            int(population),
-            int(math.ceil(population * multiplier)),
-            int(population_floor),
-        )
-
-        try:
-            max_envelope_cbgs = int(envelope_max_cbgs)
-        except (TypeError, ValueError):
-            max_envelope_cbgs = 250
-        if max_envelope_cbgs <= 0:
-            max_envelope_cbgs = 250
-
-        try:
-            max_trace_candidates = int(trace_candidate_limit)
-        except (TypeError, ValueError):
-            max_trace_candidates = 50
-        if max_trace_candidates <= 0:
-            max_trace_candidates = 50
-
-        try:
-            min_seed_capture_threshold = float(min_seed_capture)
-        except (TypeError, ValueError):
-            min_seed_capture_threshold = 0.80
-        min_seed_capture_threshold = min(1.0, max(0.0, min_seed_capture_threshold))
-
-        seed_movement_total = 0.0
-        seed_movement_always_captured = 0.0
-        seed_movement_by_cbg = {}
-        counted_seed_edges = set()
-        for seed in seed_set:
-            self_weight = float(G.nodes[seed].get('self_weight', 0) or 0)
-            seed_movement_total += self_weight
-            seed_movement_always_captured += self_weight
-            for neighbor in G.adj[seed]:
-                edge_key = frozenset((seed, neighbor))
-                if edge_key in counted_seed_edges:
-                    continue
-                counted_seed_edges.add(edge_key)
-                weight = float(G.adj[seed][neighbor].get('weight', 0) or 0)
-                seed_movement_total += weight
-                if neighbor in seed_set:
-                    seed_movement_always_captured += weight
-                else:
-                    seed_movement_by_cbg[neighbor] = (
-                        seed_movement_by_cbg.get(neighbor, 0.0) + weight
-                    )
-
-        seed_movement_captured = seed_movement_always_captured
-
-        def seed_capture_share(captured):
-            return captured / seed_movement_total if seed_movement_total > 0 else 1.0
-
-        def movement_to_cluster(candidate):
-            total = 0.0
-            for neighbor in G.adj[candidate]:
-                if neighbor in cluster_set:
-                    total += float(G.adj[candidate][neighbor].get('weight', 0) or 0)
-            return total
-
-        def cap_trace_candidates(candidates, selected_cbg, higher_score_better=True):
-            if trace_collector is None or len(candidates) <= max_trace_candidates:
-                return candidates
-
-            sorted_candidates = sorted(
-                candidates,
-                key=lambda item: (float(item.get('score', 0.0)), item.get('cbg', '')),
-                reverse=bool(higher_score_better),
-            )
-            capped = sorted_candidates[:max_trace_candidates]
-            if selected_cbg and not any(item.get('cbg') == selected_cbg for item in capped):
-                selected = next(
-                    (item for item in sorted_candidates if item.get('cbg') == selected_cbg),
-                    None,
-                )
-                if selected is not None:
-                    capped.append(selected)
-            return capped
-
-        frontier = set()
-        for seed in seed_cluster:
-            frontier.update(neighbor for neighbor in G.adj[seed] if neighbor not in cluster_set)
-
-        growth_iteration = 0
-        envelope_limited_by_cbg_cap = False
-        while (
-            (
-                population < envelope_population_target
-                or seed_capture_share(seed_movement_captured) < min_seed_capture_threshold
-            )
-            and frontier
-            and growth_iteration < max_iter
-        ):
-            if len(cluster_set) >= max_envelope_cbgs:
-                envelope_limited_by_cbg_cap = True
-                break
-
-            best = None
-            candidate_details = []
-            stale_candidates = []
-            needs_seed_capture = (
-                seed_capture_share(seed_movement_captured) < min_seed_capture_threshold
-            )
-            for candidate in sorted(frontier):
-                if candidate in cluster_set or candidate not in G:
-                    stale_candidates.append(candidate)
-                    continue
-
-                candidate_pop = get_population(candidate)
-                if candidate_pop <= 0:
-                    continue
-
-                candidate_movement = movement_to_cluster(candidate)
-                if candidate_movement <= 0:
-                    continue
-
-                movement_outside = self._movement_outside_cluster(G, candidate, cluster_set)
-                population_after = int(population + candidate_pop)
-                envelope_overshoot = max(0, population_after - envelope_population_target)
-                movement_per_person = candidate_movement / candidate_pop
-                seed_movement_gain = float(seed_movement_by_cbg.get(candidate, 0.0))
-                seed_capture_after = seed_movement_captured + seed_movement_gain
-                seed_capture_share_after = seed_capture_share(seed_capture_after)
-                seed_movement_gain_per_person = seed_movement_gain / candidate_pop
-                candidate_details.append({
-                    'cbg': candidate,
-                    'population': int(candidate_pop),
-                    'score': (
-                        float(seed_capture_share_after)
-                        if needs_seed_capture
-                        else float(candidate_movement)
-                    ),
-                    'movement_to_cluster': float(candidate_movement),
-                    'movement_to_outside': float(movement_outside),
-                    'movement_to_cluster_per_person': float(movement_per_person),
-                    'seed_movement_gain': float(seed_movement_gain),
-                    'seed_movement_gain_per_person': float(seed_movement_gain_per_person),
-                    'seed_capture_after': float(seed_capture_share_after),
-                    'population_after': int(population_after),
-                    'envelope_overshoot': int(envelope_overshoot),
-                })
-
-                if needs_seed_capture:
-                    candidate_key = (
-                        seed_movement_gain,
-                        seed_movement_gain_per_person,
-                        candidate_movement,
-                        movement_per_person,
-                        -envelope_overshoot,
-                        -movement_outside,
-                        -candidate_pop,
-                    )
-                else:
-                    candidate_key = (
-                        candidate_movement,
-                        movement_per_person,
-                        -envelope_overshoot,
-                        -movement_outside,
-                        -candidate_pop,
-                    )
-                if best is None or candidate_key > best[0]:
-                    best = (
-                        candidate_key,
-                        candidate,
-                        candidate_pop,
-                        candidate_movement,
-                        movement_outside,
-                        movement_per_person,
-                        population_after,
-                        envelope_overshoot,
-                        seed_capture_after,
-                        seed_capture_share_after,
-                        seed_movement_gain,
-                    )
-
-            for candidate in stale_candidates:
-                frontier.discard(candidate)
-
-            if best is None:
-                self.logger.info(
-                    "No positive-population mobility frontier remains for bounded prune envelope."
-                )
-                break
-
-            (
-                _,
-                selected_cbg,
-                selected_pop,
-                selected_movement,
-                selected_movement_outside,
-                selected_movement_per_person,
-                population_after,
-                selected_overshoot,
-                selected_seed_capture_after,
-                selected_seed_capture_share_after,
-                selected_seed_movement_gain,
-            ) = best
-            prev_cluster = list(cluster)
-            prev_population = int(population)
-
-            frontier.discard(selected_cbg)
-            cluster.append(selected_cbg)
-            cluster_set.add(selected_cbg)
-            population = int(population_after)
-            seed_movement_captured = selected_seed_capture_after
-            for neighbor in G.adj[selected_cbg]:
-                if neighbor not in cluster_set:
-                    frontier.add(neighbor)
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=growth_iteration,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=cap_trace_candidates(candidate_details, selected_cbg),
-                selected_cbg=selected_cbg,
-                selected_population=selected_pop,
-                cluster_after=cluster,
-                population_after=population,
-                metrics_after={
-                    'stage': 'bounded_envelope_growth',
-                    'movement_to_cluster': float(selected_movement),
-                    'movement_to_outside': float(selected_movement_outside),
-                    'movement_to_cluster_per_person': float(selected_movement_per_person),
-                    'envelope_population_target': int(envelope_population_target),
-                    'envelope_overshoot': int(selected_overshoot),
-                    'seed_movement_gain': float(selected_seed_movement_gain),
-                    'seed_capture': float(selected_seed_capture_share_after),
-                },
-            )
-
-            self.logger.info(
-                "Envelope growth iteration %d: added %s pop=%d population=%d seed_capture=%.4f movement_to_cluster=%.2f",
-                growth_iteration,
-                selected_cbg,
-                int(selected_pop),
-                int(population),
-                float(selected_seed_capture_share_after),
-                float(selected_movement),
-            )
-            growth_iteration += 1
-
-        if growth_iteration >= max_iter and population < envelope_population_target:
-            self.logger.warning(
-                "Bounded mobility-prune envelope hit max iterations before target: pop=%d target=%d",
-                int(population),
-                int(envelope_population_target),
-            )
-
-        cluster = sorted(cluster_set)
-        initial_envelope_cbg_count = len(cluster_set)
-        initial_envelope_population = int(population)
-
-        try:
-            full_stats = Helpers.calculate_movement_stats(G, cluster)
-            movement_inside = float(full_stats.get('in', 0.0))
-            movement_outside = float(full_stats.get('out', 0.0))
-        except ValueError:
-            movement_inside = 0.0
-            movement_outside = 0.0
-
-        initial_czi = (
-            movement_inside / (movement_inside + movement_outside)
-            if movement_inside + movement_outside > 0
-            else 0.0
-        )
-        initial_movement_inside = float(movement_inside)
-        initial_movement_outside = float(movement_outside)
-
-        initial_seed_capture_share = seed_capture_share(seed_movement_captured)
-
-        self.logger.info(
-            "Initial bounded mobility-prune envelope: CBGs=%d population=%d target=%d seed_capture=%.4f CZI=%.4f movement_inside=%.2f movement_boundary=%.2f",
-            len(cluster),
-            int(population),
-            int(envelope_population_target),
-            float(initial_seed_capture_share),
-            float(initial_czi),
-            float(movement_inside),
-            float(movement_outside),
-        )
-
-        def movement_effect_if_removed(candidate):
-            self_weight = float(G.nodes[candidate].get('self_weight', 0) or 0)
-            movement_to_remaining = 0.0
-            movement_to_current_outside = 0.0
-            for neighbor in G.adj[candidate]:
-                weight = float(G.adj[candidate][neighbor].get('weight', 0) or 0)
-                if neighbor in cluster_set:
-                    movement_to_remaining += weight
-                else:
-                    movement_to_current_outside += weight
-
-            movement_loss = self_weight + movement_to_remaining
-            movement_inside_after = max(0.0, movement_inside - movement_loss)
-            movement_outside_after = max(
-                0.0,
-                movement_outside - movement_to_current_outside + movement_to_remaining,
-            )
-            total_after = movement_inside_after + movement_outside_after
-            czi_after = movement_inside_after / total_after if total_after > 0 else 0.0
-            return (
-                movement_loss,
-                movement_to_current_outside,
-                movement_inside_after,
-                movement_outside_after,
-                czi_after,
-            )
-
-        def remains_seed_connected(candidate):
-            next_set = cluster_set - {candidate}
-            if not next_set:
-                return False
-            remaining_seeds = seed_set & next_set
-            if not remaining_seeds:
-                return False
-
-            seen_nodes = set()
-            stack = list(remaining_seeds)
-            while stack:
-                node = stack.pop()
-                if node in seen_nodes:
-                    continue
-                seen_nodes.add(node)
-                for neighbor in G.adj[node]:
-                    if neighbor in next_set and neighbor not in seen_nodes:
-                        stack.append(neighbor)
-            return len(seen_nodes) == len(next_set)
-
-        iteration = 0
-        stopped_by_seed_capture = False
-        while iteration < max_iter:
-            best = None
-            candidate_details = []
-            current_seed_capture_share = (
-                seed_movement_captured / seed_movement_total
-                if seed_movement_total > 0
-                else 1.0
-            )
-
-            for candidate in sorted(cluster_set):
-                if candidate in seed_set:
-                    continue
-
-                candidate_pop = population_by_cbg.get(candidate, 0)
-                if candidate_pop <= 0:
-                    continue
-
-                population_after = int(population - candidate_pop)
-
-                if not remains_seed_connected(candidate):
-                    continue
-
-                (
-                    movement_loss,
-                    movement_to_current_outside,
-                    movement_inside_after,
-                    movement_outside_after,
-                    czi_after,
-                ) = movement_effect_if_removed(candidate)
-                movement_loss_per_person = movement_loss / candidate_pop
-                seed_movement_loss = float(seed_movement_by_cbg.get(candidate, 0.0))
-                seed_capture_after = max(0.0, seed_movement_captured - seed_movement_loss)
-                seed_capture_share_after = (
-                    seed_capture_after / seed_movement_total
-                    if seed_movement_total > 0
-                    else 1.0
-                )
-                seed_movement_loss_per_person = seed_movement_loss / candidate_pop
-                would_violate_seed_capture = (
-                    seed_capture_share_after < min_seed_capture_threshold
-                    and seed_movement_loss > 1e-12
-                )
-
-                candidate_details.append({
-                    'cbg': candidate,
-                    'population': int(candidate_pop),
-                    'score': float(seed_capture_share_after),
-                    'movement_loss': float(movement_loss),
-                    'movement_loss_per_person': float(movement_loss_per_person),
-                    'movement_to_current_outside': float(movement_to_current_outside),
-                    'movement_inside_after': float(movement_inside_after),
-                    'movement_boundary_after': float(movement_outside_after),
-                    'seed_movement_loss': float(seed_movement_loss),
-                    'seed_movement_loss_per_person': float(seed_movement_loss_per_person),
-                    'seed_movement_captured_after': float(seed_capture_after),
-                    'seed_capture_after': float(seed_capture_share_after),
-                    'current_seed_capture': float(current_seed_capture_share),
-                    'czi_after': float(czi_after),
-                    'would_violate_min_seed_capture': bool(would_violate_seed_capture),
-                    'population_after': int(population_after),
-                })
-
-                if would_violate_seed_capture:
-                    continue
-
-                candidate_key = (
-                    seed_movement_loss_per_person,
-                    seed_movement_loss,
-                    movement_loss_per_person,
-                    movement_loss,
-                    -candidate_pop,
-                    candidate,
-                )
-                if best is None or candidate_key < best[0]:
-                    best = (
-                        candidate_key,
-                        candidate,
-                        candidate_pop,
-                        movement_loss,
-                        movement_inside_after,
-                        movement_outside_after,
-                        czi_after,
-                        seed_movement_loss,
-                        seed_capture_after,
-                        seed_capture_share_after,
-                    )
-
-            if best is None:
-                if candidate_details and all(
-                    bool(candidate.get('would_violate_min_seed_capture'))
-                    for candidate in candidate_details
-                ):
-                    stopped_by_seed_capture = True
-                    self.logger.info(
-                        "Stopping mobility_prune before crossing minimum seed capture %.4f",
-                        float(min_seed_capture_threshold),
-                    )
-                    break
-                self.logger.info(
-                    "No removable non-seed CBG remains without disconnecting the retained zone."
-                )
-                break
-
-            (
-                _,
-                selected_cbg,
-                removed_pop,
-                movement_loss,
-                movement_inside_after,
-                movement_outside_after,
-                czi_after,
-                seed_movement_loss,
-                seed_capture_after,
-                seed_capture_share_after,
-            ) = best
-            prev_cluster = list(cluster)
-            prev_population = int(population)
-
-            cluster_set.remove(selected_cbg)
-            cluster = [cbg for cbg in cluster if cbg != selected_cbg]
-            population -= int(removed_pop)
-            movement_inside = movement_inside_after
-            movement_outside = movement_outside_after
-            seed_movement_captured = seed_capture_after
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=iteration,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=cap_trace_candidates(
-                    candidate_details,
-                    selected_cbg,
-                    higher_score_better=True,
-                ),
-                selected_cbg=selected_cbg,
-                selected_population=removed_pop,
-                cluster_after=cluster,
-                population_after=population,
-                metrics_after={
-                    'stage': 'reverse_prune',
-                    'removed_population': int(removed_pop),
-                    'movement_loss': float(movement_loss),
-                    'movement_inside': float(movement_inside),
-                    'movement_boundary': float(movement_outside),
-                    'seed_movement_loss': float(seed_movement_loss),
-                    'seed_movement_captured': float(seed_movement_captured),
-                    'seed_capture': float(seed_capture_share_after),
-                    'czi': float(czi_after),
-                },
-                higher_score_better=True,
-            )
-
-            self.logger.info(
-                "Prune iteration %d: removed %s pop=%d population=%d seed_capture=%.4f CZI=%.4f",
-                iteration,
-                selected_cbg,
-                int(removed_pop),
-                int(population),
-                float(seed_capture_share_after),
-                float(czi_after),
-            )
-            iteration += 1
-
-        final_czi = (
-            movement_inside / (movement_inside + movement_outside)
-            if movement_inside + movement_outside > 0
-            else 0.0
-        )
-        final_seed_capture_share = (
-            seed_movement_captured / seed_movement_total
-            if seed_movement_total > 0
-            else 1.0
-        )
-        metadata = {
-            'seed_cbgs': list(seed_cluster),
-            'missing_seed_cbgs': list(missing_seed_cbgs),
-            'seed_population': int(seed_population),
-            'bounded_envelope': True,
-            'envelope_population_target': int(envelope_population_target),
-            'envelope_population_multiplier': float(multiplier),
-            'envelope_population_floor': int(population_floor),
-            'envelope_max_cbgs': int(max_envelope_cbgs),
-            'min_seed_capture': float(min_seed_capture_threshold),
-            'envelope_growth_iterations': int(growth_iteration),
-            'envelope_limited_by_cbg_cap': bool(envelope_limited_by_cbg_cap),
-            'stopped_by_seed_capture_floor': bool(stopped_by_seed_capture),
-            'initial_cbg_count': int(initial_envelope_cbg_count),
-            'initial_population': int(initial_envelope_population),
-            'initial_movement_inside': float(initial_movement_inside),
-            'initial_movement_boundary': float(initial_movement_outside),
-            'initial_czi': float(initial_czi),
-            'seed_movement_total': float(seed_movement_total),
-            'initial_seed_movement_captured': float(
-                initial_seed_capture_share * seed_movement_total
-            ),
-            'initial_seed_capture_share': float(initial_seed_capture_share),
-            'final_seed_movement_captured': float(seed_movement_captured),
-            'final_seed_capture_share': float(final_seed_capture_share),
-            'final_movement_inside': float(movement_inside),
-            'final_movement_boundary': float(movement_outside),
-            'final_czi': float(final_czi),
-            'minimum_population_used': False,
-            'legacy_min_population': int(legacy_min_population),
-            'population_target_met': True,
-            'population_reduced': int(initial_envelope_population - population),
-            'removed_cbg_count': int(initial_envelope_cbg_count - len(cluster_set)),
-        }
-
-        return cluster, int(population), metadata
-
-    def _select_candidate_nodes(self, G: nx.Graph, u0: str, candidate_limit: int):
-        if u0 not in G:
-            return [u0]
-
-        try:
-            limit = int(candidate_limit)
-        except (TypeError, ValueError):
-            limit = 120
-        limit = max(10, limit)
-
-        selected = [u0]
-        selected_set = {u0}
-        frontier = set(G.adj[u0])
-
-        while frontier and len(selected) < limit:
-            best_node = None
-            best_score = None
-
-            for node in list(frontier):
-                if node in selected_set:
-                    continue
-
-                flow_to_selected = 0.0
-                for nb in G.adj[node]:
-                    if nb in selected_set:
-                        flow_to_selected += float(G.adj[node][nb].get('weight', 0))
-
-                self_w = float(G.nodes[node].get('self_weight', 0))
-                deg = float(G.degree[node])
-                score = (flow_to_selected, self_w, deg, node)
-
-                if best_score is None or score > best_score:
-                    best_score = score
-                    best_node = node
-
-            if best_node is None:
-                break
-
-            frontier.discard(best_node)
-            if best_node in selected_set:
-                continue
-
-            selected.append(best_node)
-            selected_set.add(best_node)
-            for nb in G.adj[best_node]:
-                if nb not in selected_set:
-                    frontier.add(nb)
-
-        return selected
-
-    def _optimize_czi_subproblem(
-        self,
-        lam: float,
-        nodes,
-        edges,
-        edge_weights,
-        directed_arcs,
-        node_to_idx,
-        populations,
-        self_weights,
-        external_boundary,
-        seed_idx,
-        pop_floor: int,
-        pop_cap: int,
-        mip_rel_gap: float,
-        time_limit_sec: float
-    ):
-        n = len(nodes)
-        m = len(edges)
-        d = len(directed_arcs)
-        if n == 0:
-            return None
-
-        x_start = 0
-        y_start = n
-        z_start = n + m
-        f_start = n + (2 * m)
-        num_vars = n + (2 * m) + d
-        M = max(1, n - 1)
-
-        c = np.zeros(num_vars, dtype=float)
-        for node, idx in node_to_idx.items():
-            coeff = ((1.0 - lam) * float(self_weights.get(node, 0.0))) - (
-                lam * float(external_boundary.get(node, 0.0))
-            )
-            c[x_start + idx] = -coeff
-
-        for e_idx, w in enumerate(edge_weights):
-            c[y_start + e_idx] = -((1.0 - lam) * float(w))
-            c[z_start + e_idx] = lam * float(w)
-
-        rows = []
-        cols = []
-        vals = []
-        lb = []
-        ub = []
-        row_idx = 0
-
-        def add_row(coeffs, row_lb=-np.inf, row_ub=np.inf):
-            nonlocal row_idx
-            for col, val in coeffs.items():
-                if val == 0:
-                    continue
-                rows.append(row_idx)
-                cols.append(col)
-                vals.append(float(val))
-            lb.append(float(row_lb))
-            ub.append(float(row_ub))
-            row_idx += 1
-
-        for e_idx, (u, v) in enumerate(edges):
-            xu = x_start + node_to_idx[u]
-            xv = x_start + node_to_idx[v]
-            ye = y_start + e_idx
-            add_row({ye: 1, xu: -1}, row_ub=0)
-            add_row({ye: 1, xv: -1}, row_ub=0)
-            add_row({ye: -1, xu: 1, xv: 1}, row_ub=1)
-
-        for e_idx, (u, v) in enumerate(edges):
-            xu = x_start + node_to_idx[u]
-            xv = x_start + node_to_idx[v]
-            ze = z_start + e_idx
-            add_row({ze: 1, xu: -1, xv: 1}, row_lb=0)
-            add_row({ze: 1, xu: 1, xv: -1}, row_lb=0)
-            add_row({ze: 1, xu: -1, xv: -1}, row_ub=0)
-            add_row({ze: 1, xu: 1, xv: 1}, row_ub=2)
-
-        pop_coeff = {}
-        for node, idx in node_to_idx.items():
-            pop_coeff[x_start + idx] = float(populations.get(node, 0))
-        add_row(pop_coeff, row_lb=float(pop_floor), row_ub=float(pop_cap))
-
-        add_row({x_start + seed_idx: 1}, row_lb=1, row_ub=1)
-
-        out_arcs = {node: [] for node in nodes}
-        in_arcs = {node: [] for node in nodes}
-        for a_idx, (u, v) in enumerate(directed_arcs):
-            fa = f_start + a_idx
-            xu = x_start + node_to_idx[u]
-            xv = x_start + node_to_idx[v]
-            add_row({fa: 1, xu: -M}, row_ub=0)
-            add_row({fa: 1, xv: -M}, row_ub=0)
-            out_arcs[u].append(a_idx)
-            in_arcs[v].append(a_idx)
-
-        seed_node = nodes[seed_idx]
-
-        for node in nodes:
-            if node == seed_node:
-                continue
-
-            coeff = {x_start + node_to_idx[node]: -1}
-            for a_idx in in_arcs[node]:
-                coeff[f_start + a_idx] = coeff.get(f_start + a_idx, 0) + 1
-            for a_idx in out_arcs[node]:
-                coeff[f_start + a_idx] = coeff.get(f_start + a_idx, 0) - 1
-            add_row(coeff, row_lb=0, row_ub=0)
-
-        seed_coeff = {}
-        for a_idx in out_arcs[seed_node]:
-            seed_coeff[f_start + a_idx] = seed_coeff.get(f_start + a_idx, 0) + 1
-        for a_idx in in_arcs[seed_node]:
-            seed_coeff[f_start + a_idx] = seed_coeff.get(f_start + a_idx, 0) - 1
-        for node in nodes:
-            if node == seed_node:
-                continue
-            seed_coeff[x_start + node_to_idx[node]] = -1
-        add_row(seed_coeff, row_lb=0, row_ub=0)
-
-        A = coo_matrix((vals, (rows, cols)), shape=(row_idx, num_vars)).tocsr()
-        constraints = LinearConstraint(A, np.array(lb), np.array(ub))
-
-        lower = np.zeros(num_vars, dtype=float)
-        upper = np.concatenate([
-            np.ones(n + (2 * m), dtype=float),
-            np.full(d, float(M), dtype=float),
-        ])
-        bounds = Bounds(lower, upper)
-
-        integrality = np.zeros(num_vars, dtype=int)
-        integrality[: n + (2 * m)] = 1
-
-        options = {
-            'disp': False,
-            'mip_rel_gap': float(max(0.0, mip_rel_gap)),
-            'time_limit': float(max(1.0, time_limit_sec)),
-        }
-
-        try:
-            return milp(
-                c=c,
-                constraints=constraints,
-                integrality=integrality,
-                bounds=bounds,
-                options=options,
-            )
-        except Exception:
-            self.logger.exception("MILP subproblem failed unexpectedly")
-            return None
-
-    def czi_optimal_cap(
-        self,
-        G: nx.Graph,
-        u0: str,
-        max_pop: int,
-        candidate_limit: int = 120,
-        population_floor_ratio: float = 0.9,
-        mip_rel_gap: float = 0.02,
-        time_limit_sec: float = 20.0,
-        max_dinkelbach_iters: int = 8
-    ):
-        if u0 not in G:
-            pop = cbg_population(u0, self.config, self.logger)
-            self.logger.warning(f"Seed CBG {u0} not found in graph for czi_optimal_cap")
-            return [u0], pop
-
-        if not SCIPY_MILP_AVAILABLE:
-            self.logger.warning(
-                "SciPy MILP is unavailable; falling back to greedy_czi_balanced."
-            )
-            return self.greedy_czi_balanced(G, u0, max_pop)
-
-        try:
-            pop_cap = int(max_pop)
-        except (TypeError, ValueError):
-            pop_cap = 0
-        seed_pop = cbg_population(u0, self.config, self.logger)
-        pop_cap = max(pop_cap, seed_pop)
-
-        floor_ratio = float(population_floor_ratio)
-        floor_ratio = min(1.0, max(0.0, floor_ratio))
-        pop_floor = max(seed_pop, int(math.floor(pop_cap * floor_ratio)))
-
-        candidates = self._select_candidate_nodes(G, u0, candidate_limit)
-        node_set = set(candidates)
-        if u0 not in node_set:
-            candidates = [u0] + candidates
-            node_set = set(candidates)
-
-        populations = {}
-        self_weights = {}
-        external_boundary = {}
-        for node in candidates:
-            populations[node] = max(0, int(cbg_population(node, self.config, self.logger)))
-            self_weights[node] = float(G.nodes[node].get('self_weight', 0.0))
-
-            boundary_w = 0.0
-            for nb in G.adj[node]:
-                w = float(G.adj[node][nb].get('weight', 0))
-                if w <= 0:
-                    continue
-                if nb not in node_set:
-                    boundary_w += w
-            external_boundary[node] = boundary_w
-
-        edges = []
-        edge_weights = []
-        seen_edges = set()
-        for u in candidates:
-            for v in G.adj[u]:
-                if v not in node_set or u == v:
-                    continue
-                key = (u, v) if u < v else (v, u)
-                if key in seen_edges:
-                    continue
-                seen_edges.add(key)
-                w = float(G.adj[u][v].get('weight', 0))
-                if w <= 0:
-                    continue
-                edges.append(key)
-                edge_weights.append(w)
-
-        directed_arcs = []
-        for u, v in edges:
-            directed_arcs.append((u, v))
-            directed_arcs.append((v, u))
-
-        node_to_idx = {node: i for i, node in enumerate(candidates)}
-        seed_idx = node_to_idx[u0]
-
-        total_candidate_pop = sum(populations.values())
-        pop_floor = min(pop_floor, total_candidate_pop)
-        pop_cap = min(pop_cap, total_candidate_pop)
-        if pop_floor > pop_cap:
-            pop_floor = pop_cap
-
-        if pop_cap <= 0:
-            self.logger.warning(
-                "Candidate graph has zero population for czi_optimal_cap; falling back to greedy."
-            )
-            return self.greedy_czi_balanced(G, u0, max_pop)
-
-        self.logger.info(
-            "Starting czi_optimal_cap with %d candidates, cap=%d, floor=%d",
-            len(candidates), pop_cap, pop_floor
-        )
-
-        best_cluster = None
-        best_score = -1.0
-        lam = 0.5
-        dinkelbach_iters = max(1, int(max_dinkelbach_iters))
-
-        floor_attempts = [pop_floor]
-        while floor_attempts[-1] > seed_pop:
-            next_floor = max(seed_pop, int(math.floor(floor_attempts[-1] * 0.9)))
-            if next_floor == floor_attempts[-1]:
-                break
-            floor_attempts.append(next_floor)
-        if seed_pop not in floor_attempts:
-            floor_attempts.append(seed_pop)
-
-        for floor_idx, attempt_floor in enumerate(floor_attempts):
-            attempt_found = False
-            current_lam = lam
-
-            if floor_idx > 0:
-                self.logger.info(
-                    "Relaxing population floor to %d for czi_optimal_cap",
-                    attempt_floor
-                )
-
-            for it in range(dinkelbach_iters):
-                result = self._optimize_czi_subproblem(
-                    lam=current_lam,
-                    nodes=candidates,
-                    edges=edges,
-                    edge_weights=edge_weights,
-                    directed_arcs=directed_arcs,
-                    node_to_idx=node_to_idx,
-                    populations=populations,
-                    self_weights=self_weights,
-                    external_boundary=external_boundary,
-                    seed_idx=seed_idx,
-                    pop_floor=attempt_floor,
-                    pop_cap=pop_cap,
-                    mip_rel_gap=mip_rel_gap,
-                    time_limit_sec=time_limit_sec,
-                )
-
-                if result is None or result.x is None:
-                    self.logger.warning(
-                        "MILP solve returned no solution at Dinkelbach iteration %d (floor=%d)",
-                        it,
-                        attempt_floor
-                    )
-                    break
-
-                x = result.x[: len(candidates)]
-                chosen = [node for node, idx in node_to_idx.items() if x[idx] >= 0.5]
-                if u0 not in chosen:
-                    chosen.append(u0)
-
-                chosen_set = set(chosen)
-                chosen = [node for node in candidates if node in chosen_set]
-                if not chosen:
-                    break
-
-                try:
-                    stats = Helpers.calculate_movement_stats(G, chosen)
-                except ValueError:
-                    self.logger.warning(
-                        "Candidate solution had nodes missing from graph; falling back."
-                    )
-                    break
-
-                inside = float(stats.get('in', 0.0))
-                boundary = float(stats.get('out', 0.0))
-                denom = inside + boundary
-                czi = (inside / denom) if denom > 0 else 0.0
-                residual = inside - current_lam * denom
-
-                chosen_pop = sum(populations.get(node, 0) for node in chosen)
-                if czi > best_score or (
-                    abs(czi - best_score) < 1e-9 and best_cluster is not None and chosen_pop > sum(populations.get(node, 0) for node in best_cluster)
-                ):
-                    best_score = czi
-                    best_cluster = chosen
-
-                self.logger.info(
-                    "Dinkelbach iter %d (floor=%d): CZI=%.5f, inside=%.1f, boundary=%.1f, pop=%d, status=%s",
-                    it, attempt_floor, czi, inside, boundary, chosen_pop, str(result.status)
-                )
-
-                attempt_found = True
-                next_lam = czi
-                if abs(next_lam - current_lam) <= 1e-4 or abs(residual) <= 1e-3:
-                    current_lam = next_lam
-                    break
-                current_lam = next_lam
-
-            if attempt_found:
-                lam = current_lam
-                break
-
-        if not best_cluster:
-            self.logger.warning(
-                "czi_optimal_cap could not produce a valid MILP cluster; using greedy_czi_balanced fallback."
-            )
-            return self.greedy_czi_balanced(G, u0, max_pop)
-
-        best_population = sum(cbg_population(node, self.config, self.logger) for node in best_cluster)
-        self.logger.info(
-            "czi_optimal_cap selected %d CBGs, population=%d, CZI=%.5f",
-            len(best_cluster), best_population, best_score
-        )
-        return best_cluster, best_population
-
-    def greedy_czi_balanced(self, G: nx.Graph, u0: str, min_pop: int,
-                            alpha: float = 0.75, overshoot_penalty: float = 0.25,
-                            distance_penalty_weight: float = 0.02,
-                            distance_scale_km: float = 20.0,
-                            cbg_centers=None,
-                            trace_collector=None):
-        self.logger.info(f"Starting greedy_czi_balanced algorithm with seed CBG {u0}")
-        population = cbg_population(u0, self.config, self.logger)
-        cluster = [u0]
-        cluster_set = {u0}
-        cbg_centers = cbg_centers or {}
-        seed_center = cbg_centers.get(u0)
-
-        if u0 not in G:
-            self.logger.warning(f"Seed CBG {u0} not found in graph")
-            return cluster, population
-
-        base_stats = Helpers.calculate_movement_stats(G, cluster)
-        movement_in = float(base_stats.get('in', 0))
-        movement_out = float(base_stats.get('out', 0))
-        surround = set(G.adj[u0])
-
-        itr = 0
-        while population < min_pop:
-            if not surround:
-                self.logger.warning(
-                    f"No adjacent CBGs found after {itr} iterations. Cannot reach target population."
-                )
-                break
-
-            best = None
-            candidate_details = []
-
-            for candidate in list(surround):
-                if candidate in cluster_set or candidate not in G:
-                    continue
-
-                cand_pop = cbg_population(candidate, self.config, self.logger)
-                if cand_pop <= 0:
-                    continue
-
-                self_weight = float(G.nodes[candidate].get('self_weight', 0))
-                in_to_cluster = 0.0
-                out_to_outside = 0.0
-                for neighbor in G.adj[candidate]:
-                    weight = float(G.adj[candidate][neighbor].get('weight', 0))
-                    if weight <= 0:
-                        continue
-                    if neighbor in cluster_set:
-                        in_to_cluster += weight
-                    else:
-                        out_to_outside += weight
-
-                inside_after = movement_in + self_weight + in_to_cluster
-                boundary_after = movement_out - in_to_cluster + out_to_outside
-                if boundary_after < 0 and abs(boundary_after) < 1e-9:
-                    boundary_after = 0.0
-                total_after = inside_after + boundary_after
-                czi_after = (inside_after / total_after) if total_after > 0 else 0.0
-
-                distance_penalty = 0.0
-                if seed_center and distance_scale_km > 0:
-                    cand_center = cbg_centers.get(candidate)
-                    if cand_center:
-                        dist_km = distance(
-                            seed_center[0], seed_center[1],
-                            cand_center[0], cand_center[1]
-                        )
-                        distance_penalty = dist_km / (dist_km + distance_scale_km)
-
-                score = (
-                    czi_after
-                    - distance_penalty_weight * distance_penalty
-                )
-
-                candidate_tuple = (
-                    score,
-                    czi_after,
-                    -distance_penalty,
-                    in_to_cluster,
-                    candidate,
-                    cand_pop,
-                    inside_after,
-                    boundary_after
-                )
-                candidate_details.append({
-                    'cbg': candidate,
-                    'population': int(cand_pop),
-                    'score': float(score),
-                    'movement_to_cluster': float(in_to_cluster),
-                    'movement_to_outside': float(out_to_outside),
-                    'czi_after': float(czi_after),
-                    'distance_penalty': float(distance_penalty),
-                    'movement_inside_after': float(inside_after),
-                    'movement_boundary_after': float(boundary_after),
-                })
-                if best is None or candidate_tuple[:4] > best[:4]:
-                    best = candidate_tuple
-
-            if best is None:
-                self.logger.warning(
-                    f"No valid candidate CBGs found after {itr} iterations. Cannot reach target population."
-                )
-                break
-
-            _, czi_after, _, _, best_cbg, best_pop, inside_after, boundary_after = best
-            prev_cluster = list(cluster)
-            prev_population = population
-            cluster.append(best_cbg)
-            cluster_set.add(best_cbg)
-            population += best_pop
-            movement_in = inside_after
-            movement_out = boundary_after
-
-            self._record_trace_step(
-                trace_collector,
-                iteration=itr,
-                cluster_before=prev_cluster,
-                population_before=prev_population,
-                candidates=candidate_details,
-                selected_cbg=best_cbg,
-                selected_population=best_pop,
-                cluster_after=cluster,
-                population_after=population,
-                metrics_after={
-                    'czi': float(czi_after),
-                    'movement_inside': float(movement_in),
-                    'movement_boundary': float(movement_out),
-                }
-            )
-
-            surround.remove(best_cbg)
-            surround.update([j for j in G.adj[best_cbg] if j not in cluster_set])
-
-            self.logger.info(
-                f"Iteration {itr}: Added CBG {best_cbg} with pop {best_pop}. "
-                f"New total: {population}. CZI: {czi_after:.4f}"
-            )
-
-            itr += 1
-            if itr > 500:
-                self.logger.warning("Max iterations exceeded (500). Cannot reach target population.")
-                break
-
-        return cluster, population
